requirements.txt.txt
fastapi 
uvicorn 
python-dotenv 
twilio 
elevenlabs 
websockets
mangum
supabase
pydantic_settings